// src/lib/api/generated/models/objectDataPatchDocumentItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.3
 */
import type { AddOperation } from './addOperation';
import type { CopyOperation } from './copyOperation';
import type { MoveOperation } from './moveOperation';
import type { RemoveOperation } from './removeOperation';
import type { ReplaceOperation } from './replaceOperation';
import type { TestOperation } from './testOperation';

/**
 * JSON Patch single patch operation
 */
export type ObjectDataPatchDocumentItem = AddOperation & {
  op: 'add';
} | RemoveOperation & {
  op: 'remove';
} | ReplaceOperation & {
  op: 'replace';
} | MoveOperation & {
  op: 'move';
} | CopyOperation & {
  op: 'copy';
} | TestOperation & {
  op: 'test';
};